refactor(ai-engine): resolve #1744 — split java_analyzer/ monolith cluster into focused modules
Split the three large files in agents/java_analyzer/ into focused submodules
while preserving the public API via re-export shims and mixin composition:

tools.py (40K -> 4K shim):
  * _base_tool.py        - shared _BaseJavaAnalyzerTool
  * structure_tools.py   - mod structure & metadata tools
  * feature_tools.py     - feature & asset identification tools
  * dependency_tools.py  - dependency analysis tool
  * complexity_tools.py  - LLM complexity analysis tool

feature_extractor.py (45K -> 23K facade):
  * _constants.py             - availability flags, parser handles, limits
  * _helpers.py               - registry-name helpers
  * _parsing_mixin.py         - tree-sitter parse + AST text helpers
  * _goal_extraction_mixin.py - entity/goal selector extraction
  * _analysis_mixin.py        - metadata/dependency/reflection analysis

java_analyzer.py (30K -> 12K orchestrator):
  * _jar_analysis_mixin.py - JAR/source/registry analysis

The original modules remain as backwards-compatibility shims re-exporting
every public name (JavaAnalyzerTools facade, FeatureExtractor, the typed
BaseTool subclasses, Pydantic schemas, availability flags), so existing
imports are unchanged. All java/analyzer/feature tests pass (438 passed).

Closes #1744
